fix(region): disk fail notify
