+++ pkg/compute/tasks/dnszone/dns_record_set_enabled_task.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package disk
+package dnszone
 
 import (
 	"context"
